feat: use recursive component
