+++ packages/api-reference/src/components/Content/Operations/index.ts
@@ -1,1 +1,1 @@
-export { default as OperationsAndWebhooks } from './OperationsAndWebhooks.vue'
+export { default as TraversedEntryContainer } from './TraversedEntryContainer.vue'
