+++ packages/api-reference/src/components/Lazy/index.ts
@@ -1,3 +1,2 @@
 export { default as Lazy } from './Lazy.vue'
-export { default as Loading } from './Loading.vue'
 export { lazyBus } from './lazyBus'
